Split memory tuning advisor panel
Move JVM and Kubernetes sizing tools into a dedicated Tuning Advisor panel while keeping Memory focused on live metrics.
