+++ bootui-autoconfigure/src/main/java/io/github/jdubois/bootui/autoconfigure/web/MemoryController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/bootui/api/memory")
+@RequestMapping({"/bootui/api/memory", "/bootui/api/tuning-advisor"})
 public class MemoryController {
 
     private final MemoryCalculator calculator;
